code_examples/Python/app_debugger/test_client/requirements.txt
suds-jurko
clfpy
